Move helper functions into `helpers.py`
